app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
psycopg2-binary
alembic
python-dotenv
pydantic-settings
python-jose
redis
passlib==1.7.4
bcrypt==3.2.0
